# project_docs/generated/openapi-v1.json
{
  "components": {
    "schemas": {
      "ApproveConfirmationRequest": {
        "properties": {
          "fields": {
            "additionalProperties": true,
            "title": "Fields",
            "type": "object"
          }
        },
        "title": "ApproveConfirmationRequest",
        "type": "object"
      },
...
      "ConfirmationData": {
        "properties": {
          "approved_by_actor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By Actor Id"
          },
          "confirmation_id": {
            "title": "Confirmation Id",
            "type": "string"
          },
          "confirmation_type": {
            "title": "Confirmation Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fields": {
            "additionalProperties": true,
            "title": "Fields",
            "type": "object"
          },
          "requested_by_employee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Employee Id"
          },
          "resolution_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Payload"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_run_id": {
            "title": "Task Run Id",
            "type": "string"
          }
        },
        "required": [
          "confirmation_id",
          "session_id",
          "task_run_id",
          "confirmation_type",
          "status",
          "fields",
          "requested_by_employee_id",
          "resolution_payload",
          "approved_by_actor_id",
          "created_at",
          "resolved_at"
        ],
        "title": "ConfirmationData",
        "type": "object"
      },
...
      "DataEnvelope_ConfirmationData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ConfirmationData"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataEnvelope[ConfirmationData]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListConfirmationsMeta": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count"
        ],
        "title": "ListConfirmationsMeta",
        "type": "object"
      },
      "ListConfirmationsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ConfirmationData"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ListConfirmationsMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ListConfirmationsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Store Manager Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/confirmations": {
      "get": {
        "operationId": "get_confirmations_api_v1_confirmations_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "pending",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListConfirmationsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Confirmations",
        "tags": [
          "confirmations"
        ]
      }
    },
    "/api/v1/confirmations/{confirmation_id}/approve": {
      "post": {
        "operationId": "post_approve_confirmation_api_v1_confirmations__confirmation_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "confirmation_id",
            "required": true,
            "schema": {
              "title": "Confirmation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ConfirmationData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Approve Confirmation",
        "tags": [
          "confirmations"
        ]
      }
    },
    "/api/v1/confirmations/{confirmation_id}/reject": {
      "post": {
        "operationId": "post_reject_confirmation_api_v1_confirmations__confirmation_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "confirmation_id",
            "required": true,
            "schema": {
              "title": "Confirmation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ConfirmationData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Reject Confirmation",
        "tags": [
          "confirmations"
        ]
      }
    },
...
  }
}